Add: if not exist zshenv

+++ .config/zsh/install.sh
@@ -10,6 +10,10 @@ export ZDOTDIR='${DOTFILES_ZSH}'
 export XDG_CONFIG_HOME='${DOTFILES}/.config'
 export XDG_DATA_HOME='${DOTFILES}/.local/share'
 EOF
-  mv "${HOME}/.zshenv" "${HOME}/zshenv.bak"
-  cat "${XDG_ENV}" "/tmp/zshenv" > "${HOME}/.zshenv"
+  if [ -f "${HOME}/.zshenv" ];then
+    mv "${HOME}/.zshenv" "${HOME}/zshenv.bak"
+    cat "${XDG_ENV}" "/tmp/zshenv" > "${HOME}/.zshenv"
+  else
+    cp "${XDG_ENV}" "${HOME}/.zshenv"
+  fi
 fi
